% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{autoplot.elife_hazard}
\alias{autoplot.elife_hazard}
\title{Create a ggplot object for hazards}
\usage{
autoplot.elife_hazard(object, ...)
}
\arguments{
\item{object}{object of class \code{elife_hazard}}

\item{...}{additional parameters, currently ignored by the function}
}
\value{
a \code{ggplot} object
}
\description{
If the package \code{ggplot2} is installed,
create a \code{ggplot} object.
}
\keyword{internal}
