% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{autoplot.elife_northropcoleman}
\alias{autoplot.elife_northropcoleman}
\title{Create a ggplot object for the piece-wise generalized Pareto diagnostics}
\usage{
autoplot.elife_northropcoleman(x, plot = TRUE, ...)
}
\arguments{
\item{...}{additional parameters, passed to plot}

\item{object}{object of class \code{elife_northropcoleman}}
}
\value{
a \code{ggplot} object
}
\description{
If the package \code{ggplot2} is installed,
create a \code{ggplot} object.
}
\keyword{internal}
