% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{autoplot.elife_par}
\alias{autoplot.elife_par}
\title{Create a ggplot object for fitted parametric models}
\usage{
autoplot.elife_par(object, ...)
}
\arguments{
\item{object}{object of class \code{elife_par}}

\item{...}{additional graphical parameters}
}
\value{
a \code{ggplot} object
}
\description{
If the package \code{ggplot2} is installed,
the function creates a \code{ggplot} object.
}
\keyword{internal}
