% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshstab.R
\name{autoplot.elife_tstab}
\alias{autoplot.elife_tstab}
\title{Create a ggplot object for threshold stability plots}
\usage{
autoplot.elife_tstab(object, ...)
}
\arguments{
\item{object}{object of class \code{elife_tstab}}

\item{...}{additional parameters for labels, etc.}
}
\value{
a \code{ggplot} object
}
\description{
If the package \code{ggplot2} is installed,
create a \code{ggplot} object.
}
\keyword{internal}
