% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{boxcox_transfo}
\alias{boxcox_transfo}
\title{Box-Cox transformation function}
\usage{
boxcox_transfo(x, lambda = rep(1, length(x)))
}
\arguments{
\item{x}{vector of arguments}

\item{lambda}{vector of Box-Cox parameters}
}
\value{
a vector of the same length as \code{x}
}
\description{
Given a vector of parameters, apply the Box-Cox transformation.
}
\keyword{internal}
