% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{check_arguments}
\alias{check_arguments}
\title{Check default arguments}
\usage{
check_arguments(func, call, arguments = NULL)
}
\arguments{
\item{func}{function whose parameters are to be superseded}

\item{call}{user call, obtained from \code{match.call(expand.dots = FALSE)}}

\item{arguments}{named list with arguments}
}
\value{
a named list with all arguments
}
\description{
Check arguments and override default values.
If a named list, \code{arguments}, is provided by the user,
it will override any default value.
If one of the argument is provided directly,
it will take precedence over the values in \code{arguments}, provided it is not a default value.
}
\keyword{internal}
