% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{check_elife_dist}
\alias{check_elife_dist}
\title{Check parameters of extended lifetime distributions}
\usage{
check_elife_dist(
  rate,
  scale,
  shape,
  family = c("exp", "gp", "weibull", "gomp", "gompmake", "extgp", "extweibull", "perks",
    "beard", "perksmake", "beardmake")
)
}
\arguments{
\item{rate}{rate parameter(s); for models with Makeham component, the last entry should be part of the rate vector}

\item{scale}{scale parameter}

\item{shape}{vector of shape parameter(s).}

\item{family}{string indicating the parametric model, one of \code{exp}, \code{gp}, \code{gomp}, \code{gompmake}, \code{weibull}, \code{extgp}, \code{extweibull}, \code{perks}, \code{perksmake}, \code{beard} and \code{beardmake}}
}
\description{
Check parameters of extended lifetime distributions
}
\keyword{internal}
