% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dbeard}
\alias{dbeard}
\title{Density function of the Beard distribution}
\usage{
dbeard(x, rate = 1, shape1 = 1, shape2 = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape1}{shape parameter (\eqn{\alpha})}

\item{shape2}{shape parameter (\eqn{\beta})}

\item{log}{logical; if \code{FALSE} (default), return the density, else the log likelihood of the individual observations.}
}
\value{
a vector of (log)-density.
}
\description{
Density function of the Beard distribution
}
\keyword{internal}
