% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dbeardmake}
\alias{dbeardmake}
\title{Density function of the Beard-Makeham distribution}
\usage{
dbeardmake(x, rate = rate, shape1 = 1, shape2 = 1, lambda = 0, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{shape parameter (\eqn{\nu})}

\item{shape1}{shape parameter (\eqn{\alpha})}

\item{shape2}{shape parameter (\eqn{\beta})}

\item{lambda}{exponential rate of the Makeham component}

\item{log}{logical; if \code{FALSE} (default), return the density, else the log likelihood of the individual observations.}
}
\value{
a vector of (log)-density.
}
\description{
Density function of the Beard-Makeham distribution
}
\keyword{internal}
