% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dextgp}
\alias{dextgp}
\title{Density function of the extended generalized Pareto distribution}
\usage{
dextgp(x, scale = 1, shape1, shape2, log = FALSE)
}
\arguments{
\item{scale}{scale parameter, strictly positive.}

\item{shape1}{positive shape parameter \eqn{\beta}; model defaults to generalized Pareto when it equals zero.}

\item{shape2}{shape parameter \eqn{\gamma}; model reduces to Gompertz when \code{shape2=0}.}

\item{log}{logical; if \code{TRUE}, return the log-density}
}
\value{
a vector of (log)-density of the same length as \code{x}
}
\description{
Density function of the extended generalized Pareto distribution
}
\keyword{internal}
