% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dextweibull}
\alias{dextweibull}
\title{Density function of the extended Weibull distribution}
\usage{
dextweibull(x, scale = 1, shape1 = 0, shape2 = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{scale}{scale parameter, strictly positive.}

\item{shape1}{shape parameter of the generalized Pareto component.}

\item{shape2}{shape parameter of the Weibull component.}

\item{log}{logical; if \code{FALSE} (default), return the density, else the log likelihood of the individual observations.}
}
\value{
a vector of (log)-density.
}
\description{
Density function of the extended Weibull distribution
}
\keyword{internal}
