% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgomp}
\alias{dgomp}
\title{Density function of the Gompertz distribution}
\usage{
dgomp(x, scale = 1, shape, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{scale}{positive scale parameter}

\item{shape}{non-negative shape parameter}

\item{log}{logical; if \code{TRUE}, return the log density}
}
\description{
Density function of the Gompertz distribution
}
\keyword{internal}
