% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dgompmake}
\alias{dgompmake}
\title{Density function of the Gompertz-Makeham distribution}
\usage{
dgompmake(x, scale = 1, shape = 1, lambda = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{scale}{scale parameter, strictly positive.}

\item{shape}{shape parameter.}

\item{lambda}{exponential rate}
}
\value{
vector of density
}
\description{
Density function of the Gompertz-Makeham distribution
}
\keyword{internal}
