% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{dgpd}
\alias{dgpd}
\title{Density function of the generalized Pareto distribution}
\usage{
dgpd(x, loc = 0, scale = 1, shape = 0, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{loc}{location parameter.}

\item{scale}{scale parameter, strictly positive.}

\item{shape}{shape parameter.}

\item{log}{logical; if \code{FALSE} (default), return the density, else the log likelihood of the individual observations.}
}
\value{
a vector of (log)-density.
}
\description{
Density function of the generalized Pareto distribution
}
\keyword{internal}
