% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.censTruncLimitsDtrunc}
\alias{.censTruncLimitsDtrunc}
\title{Identification sets for double truncated data}
\usage{
.censTruncLimitsDtrunc(tsets, lcens, rcens, ltrunc, rtrunc, trunc, cens)
}
\arguments{
\item{tsets}{Turnbull's sets}

\item{lcens}{numeric vector of left censoring}

\item{rcens}{numeric vector of right censoring}

\item{ltrunc}{numeric matrix of left truncation}

\item{rtrunc}{numeric matrix of right truncation}

\item{trunc}{logical are observation truncated?}
}
\value{
a matrix with the bounds of the intervals for Turnbull sets
}
\description{
Identification sets for double truncated data
}
\keyword{internal}
