% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{.check_surv}
\alias{.check_surv}
\title{Check survival output}
\usage{
.check_surv(
  time,
  time2 = NULL,
  event = NULL,
  type = c("right", "left", "interval", "interval2")
)
}
\arguments{
\item{time}{excess time of the event of follow-up time, depending on the value of event}

\item{time2}{ending excess time of the interval for interval censored data only.}

\item{event}{status indicator, normally 0=alive, 1=dead. Other choices are \code{TRUE}/\code{FALSE} (\code{TRUE} for death).
For interval censored data, the status indicator is 0=right censored, 1=event at time, 2=left censored, 3=interval censored.
Although unusual, the event indicator can be omitted, in which case all subjects are assumed to have experienced an event.}

\item{type}{character string specifying the type of censoring. Possible values are "\code{right}", "\code{left}", "\code{interval}", "\code{interval2}".}
}
\value{
a list with transformed inputs or an error
}
\description{
Check survival output
}
\keyword{internal}
