% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonparametric.R
\name{.wecdf}
\alias{.wecdf}
\title{Weighted empirical distribution function}
\usage{
.wecdf(x, w, type = c("dist", "surv"))
}
\arguments{
\item{x}{vector of length \code{n} of values}

\item{w}{vector of weights of length \code{n}}

\item{type}{string, one of distribution function (\code{dist}) or survival function (\code{surv})}
}
\description{
Weighted empirical distribution function
}
\author{
Adapted from spatstat (c) Adrian Baddeley and Rolf Turner
}
\keyword{internal}
