% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dperks}
\alias{dperks}
\title{Density function of the Perks distribution}
\usage{
dperks(x, rate = 1, shape = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape}{shape parameter (\eqn{\alpha})}

\item{log}{logical; if \code{FALSE} (default), return the density, else the log likelihood of the individual observations.}
}
\value{
a vector of (log)-density.
}
\description{
Density function of the Perks distribution
}
\keyword{internal}
