% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dperksmake}
\alias{dperksmake}
\title{Density function of the Perks-Makeham distribution}
\usage{
dperksmake(x, rate = 1, shape = 1, lambda = 0, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape}{shape parameter (\eqn{\alpha})}

\item{lambda}{exponential rate of the Makeham component}

\item{log}{logical; if \code{FALSE} (default), return the density, else the log likelihood of the individual observations.}
}
\value{
a vector of (log)-density.
}
\description{
Density function of the Perks-Makeham distribution
}
\keyword{internal}
