% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{englandwales}
\alias{englandwales}
\title{England and Wales age at death}
\format{
A data frame with 4032 rows and 10 variables:
\describe{
\item{ndays}{survival time (in days)}
\item{ageyear}{floor of maximum age reached (in years)}
\item{gender}{factor indicating gender of individual, either \code{female} or \code{male}}
\item{valid}{factor for the quality flag; \code{yes} for validated records, \code{no} for unchecked records}
\item{ddate}{death date (\code{Date})}
\item{bdate}{birth date (\code{Date})}
\item{ltrunc1}{first lower truncation bound (in days) corresponding to the lowest possible age achieved for inclusion as semisupercentenarian or supercentenarian}
\item{rtrunc1}{first upper truncation bound (in days), corresponding to the maximum age reachable by the end of the sampling frame time window}
\item{ltrunc2}{second lower truncation bound (in days) corresponding to 110 years, for semisupercentenarians who could have reached 110 before 2021}
\item{rtrunc2}{second upper truncation bound (in days) corresponding to the maximum age by 2021 of semisupercentenarians who could have reached 110 before then}
}
}
\source{
International Database on Longevity (IDL), 2021. Hosted by the French Institute for Demographic Studies (Ined) at \url{https://www.supercentenarians.org/en/}.
Data accessed on December 1st, 2023
}
\usage{
englandwales
}
\description{
This data frame contains information about 4032
Welsh and English individuals who died at age ranging from
105 to 110 (so-called semisupercentenarians) between
December 28th, 1999 and December 31st, 2014, as well
as all 179 supercentenarians (aged 110 and above) who died
between January 1st, 1968 and December 31st, 2020.
}
\details{
All records for people who died at age 109 and all men,
plus a stratified sample of the women were validated
by the General Register Office (GRO). Observations are
doubly interval truncated: some individuals could have
been observed over distinct intervals if they died between
105 and 110 before 2014, but could have reached 110 before
the end of the collection period.


The referenced technical report from ONS describes the validation
procedure in more details and includes (approximate) sampling weights
for the validation of the sample of women who died age 105-108.

The IDL data are licensed under CC BY 4.0
License; please refer to the IDL database to download the latest available data.
}
\references{
Office for National Statistics (2016). Accuracy of official high-age
population estimates, in England and Wales: an evaluation. Technical report,
\url{https://www.ons.gov.uk/peoplepopulationandcommunity/birthsdeathsandmarriages/ageing/methodologies/accuracyofofficialhighagepopulationestimatesinenglandandwalesanevaluation}

International Database on Longevity (2021). French Institute for Demographic Studies (Ined) (host).
}
\keyword{datasets}
