% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ewsim}
\alias{ewsim}
\title{England and Wales simulated supercentenarian data}
\format{
A data frame with 179 rows and 3 variables:
\describe{
\item{time}{survival time above 110 (in years)}
\item{ltrunc}{minimum age above 110 (in years), or zero};
\item{rtrunc}{maximum age (in years) an individual could have reached by the end of the time frame}
}
}
\usage{
ewsim
}
\description{
This data frame contains information about 179 fake records mimicking
Welsh and English who died age 110 and above
}
\keyword{datasets}
