% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npecdf.R
\name{expit}
\alias{expit}
\title{Inverse logistic link function}
\usage{
expit(x)
}
\arguments{
\item{x}{vector of real values}
}
\value{
probability vector
}
\description{
Inverse logistic link function
}
\keyword{internal}
