% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{hexp}
\alias{hexp}
\title{Hazard function of the exponential distribution}
\usage{
hexp(x, rate = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{rate}{rate vector of rates}

\item{log}{logical; if \code{FALSE} (default), return the log hazard}
}
\value{
a vector of (log)-hazard.
}
\description{
Hazard function of the exponential distribution
}
\keyword{internal}
