% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fit_elife}
\alias{fit_elife}
\title{Maximum likelihood estimation of parametric models for excess lifetime}
\usage{
fit_elife(
  dat,
  thresh,
  ltrunc,
  rtrunc,
  rcens,
  type = c("none", "ltrc", "ltrt"),
  family = c("exp", "gp", "weibull", "gomp", "extgp"),
  weights = NULL
)
}
\arguments{
\item{thresh}{a vector of thresholds}

\item{family}{string, one of \code{exp}, \code{gp}, \code{gomp} or \code{ext} for exponential, generalized Pareto, Gompertz or extended family, respectively.}

\item{weights}{vector of weights, defaults to one for each observation.}
}
\value{
a list containing
}
\description{
This function performs constrained optimization to find the maximum likelihood estimator
for a range of parametric models suitable for the analysis of longevity data.
These can be estimated simultaneously for a range of thresholds.
}
