% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{fitrange_elife}
\alias{fitrange_elife}
\title{Maximum likelihood estimation of parametric models for excess lifetime}
\usage{
fitrange_elife(
  time,
  time2 = NULL,
  event = NULL,
  ltrunc = NULL,
  rtrunc = NULL,
  thresh,
  type = c("right", "left", "interval", "interval2"),
  family = c("exp", "gp", "weibull", "gomp", "gompmake", "extgp"),
  weights = NULL
)
}
\arguments{
\item{time}{excess time of the event of follow-up time, depending on the value of event}

\item{time2}{ending excess time of the interval for interval censored data only.}

\item{event}{status indicator, normally 0=alive, 1=dead. Other choices are \code{TRUE}/\code{FALSE} (\code{TRUE} for death).
For interval censored data, the status indicator is 0=right censored, 1=event at time, 2=left censored, 3=interval censored.
Although unusual, the event indicator can be omitted, in which case all subjects are assumed to have an event.}

\item{ltrunc}{lower truncation limit, default to \code{NULL}}

\item{rtrunc}{upper truncation limit, default to \code{NULL}}

\item{thresh}{a vector of thresholds}

\item{type}{character string specifying the type of censoring. Possible values are "\code{right}", "\code{left}", "\code{interval}", "\code{interval2}".}

\item{family}{string, one of \code{exp}, \code{gp}, \code{gomp} or \code{ext} for exponential, generalized Pareto, Gompertz or extended family, respectively.}

\item{weights}{vector of weights, defaults to one for each observation.}
}
\value{
a list containing
}
\description{
This function performs constrained optimization to find the maximum likelihood estimator
for a range of parametric models suitable for the analysis of longevity data.
These can be estimated simultaneously for a range of thresholds.
}
