% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard_elife}
\alias{hazard_elife}
\title{Profile likelihood for hazard}
\usage{
hazard_elife(
  x,
  dat,
  thresh = 0,
  type = c("none", "ltrt", "ltrc"),
  rcens,
  ltrunc,
  rtrunc,
  family = c("exp", "gp", "gomp", "weibull", "extgp"),
  weights = rep(1, length(dat)),
  level = 0.95,
  psi = NULL,
  plot = FALSE
)
}
\arguments{
\item{x}{value of the threshold exceedance at which to estimate the hazard}

\item{dat}{vector of threshold exceedances}

\item{thresh}{threshold}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}

\item{weights}{weights for observations}

\item{plot}{logical; if true, display the profile log-likelihood. Default to \code{FALSE}.}
}
\value{
an invisible object of class \code{elife_hazard} containing information about the profile likelihood
}
\description{
This function computes the hazard for different \code{elife} parametric
models with profile-likelihood based confidence intervals.
It is also used to provide local hazard plots at varying thresholds.
}
\examples{
n <- 2500
dat <- rdtrunc_elife(n = n, scale = 2,
family = "gp", shape = 0.1,
lower = ltrunc <- runif(n),
upper = rtrunc <- (5 + runif(n)))
hazard_elife(x = 2, dat = dat,
 ltrunc = ltrunc, rtrunc = rtrunc,
 type = "ltrt", family = "gp")
}
