% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{hweibull}
\alias{hweibull}
\title{Hazard function of the Weibull distribution}
\usage{
hweibull(x, shape, scale = 1, log = FALSE)
}
\arguments{
\item{x}{vector of quantiles}

\item{shape}{shape parameter}

\item{scale}{scale parameter, default to 1}

\item{log}{logical; if \code{TRUE}, returns the log hazard}
}
\value{
a vector of (log)-hazard
}
\description{
Hazard function of the Weibull distribution
}
\keyword{internal}
