% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{japanese2}
\alias{japanese2}
\title{Japanese survival data (2)}
\format{
A data frame with 216 rows and 4 variables:
\describe{
\item{age}{integer, age (to the smallest year) at death (in years)}
\item{bcohort}{factor, birth cohort}
\item{count}{integer, number of death for cohort at given age}
\item{gender}{factor, the gender of the individuals; either \code{male} or \code{female}}
}
}
\source{
Table 10.3
}
\usage{
japanese2
}
\description{
This data frame is extracted from Table 10.3 from Chapter 10, "Centenarians and Supercentenarians in Japan", in the Monograph Exceptional lifespans. The data were constructed by the extinct cohort method and are stratified by age cohort (five year group, except 1899-1900) and by sex. Note that the family registry system (KOSEKI), introduced in 1872, was standardized in 1886.
}
\references{
Saito, Yasuhiko and Futoshi Ishii, and Jean-Marie Robine (2021). \emph{Centenarians and Supercentenarians in Japan}. In \emph{Exceptional lifespans}, Maier, H., Jeune, B., Vaupel, J. W. (Eds.), Demographic research monographs 17 VII, pp. 125-145. Cham, Springer.
}
\keyword{datasets}
