% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npecdf.R
\name{logit}
\alias{logit}
\title{Logistic link function}
\usage{
logit(x)
}
\arguments{
\item{x}{vector of probabilities}
}
\value{
vector of real values
}
\description{
Logistic link function
}
\keyword{internal}
