% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{nc_score_test}
\alias{nc_score_test}
\title{Score test of Northrop and Coleman}
\usage{
nc_score_test(
  dat,
  thresh,
  ltrunc = NULL,
  rtrunc = NULL,
  rcens = NULL,
  type = c("none", "ltrc", "ltrt"),
  weights = rep(1, length(dat))
)
}
\arguments{
\item{dat}{vector of threshold exceedances}

\item{thresh}{a vector of thresholds}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{weights}{weights for observations}
}
\value{
a data frame with the following variables:
\itemize{
\item{\code{thresh}: }{threshold for the generalized Pareto}
\item{\code{nexc}: }{number of exceedances}
\item{\code{score}: }{score statistic}
\item{\code{df}: }{degrees of freedom}
\item{\code{pval}: }{the p-value obtained from the asymptotic chi-square approximation.}
}
}
\description{
This function computes the score test
with the piecewise generalized Pareto distribution
under the null hypothesis that the generalized Pareto
with a single shape parameter is an adequate simplification.
The test statistic is calculated using the observed information
matrix; both hessian and score vector are obtained through numerical differentiation.
}
\details{
The reference distribution is chi-square
}
