% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{nll_elife}
\alias{nll_elife}
\title{Likelihood for left-truncated and right-censored or interval-truncated data}
\usage{
nll_elife(
  par,
  dat,
  thresh = 0,
  type = c("none", "ltrt", "ltrc"),
  rcens,
  ltrunc,
  rtrunc,
  family = c("exp", "gp", "gomp", "weibull", "extgp"),
  weights = rep(1, length(dat))
)
}
\arguments{
\item{par}{vector of parameters}

\item{dat}{vector of threshold exceedances}

\item{thresh}{threshold}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}

\item{weights}{weights for observations}
}
\value{
log-likelihood value
}
\description{
Computes the log-likelihood for various parametric models suitable for threshold
exceedances
}
