% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsurv.R
\name{np.interv.mle}
\alias{np.interv.mle}
\title{Nonparametric estimation of the survival function}
\usage{
np.interv.mle(
  dat,
  thresh,
  delta,
  rcens = NULL,
  ltrunc = NULL,
  rtrunc = NULL,
  tol = 1e-12
)
}
\arguments{
\item{dat}{vector of exceedances of length \code{n}, i.e., observations that are larger than the thresh}

\item{thresh}{double thresh}

\item{delta}{width of the intervals}

\item{rcens}{logical vector of length \code{n}, where \code{rcens} is \code{TRUE} for right-censored observation and \code{FALSE} otherwise.}

\item{ltrunc}{vector of size \code{n} of lower truncation time}

\item{rtrunc}{vector of size \code{n} of upper truncation time}

\item{tol}{double, relative tolerance for convergence of the EM algorithm}
}
\value{
an object of class \code{stepfun}
}
\description{
The survival function is obtained through the EM algorithm
described in Turnbull (1978) for left-truncated right-censored
or else doubly truncated observations; censoring is assumed to be
non-informative. The exceedances \eqn{y_i-u}
are separated into discretized intervals of the form
\deqn{p_j = \Pr[(j-1)\delta \leq Y-u < j\delta], \quad j=1, \ldots, J}
where \eqn{J} is chosen such that \eqn{\max(Y_j) - u < J \delta}.
}
\details{
The unknown parameters of the model are \eqn{p_j (j=1, \ldots, J)}
subject to the constraint that \eqn{\sum_{j=1}^J p_j=1}.
}
\examples{
set.seed(2021)
n <- 100L
dat <- evd::rgpd(n = n, scale = 15, shape = -0.1)
ltrunc <- rep(0, n)
rtrunc <- rpois(n = n, lambda = 50) + dat
rcens <- NULL
thresh <- 0
delta <- 2
npi <- np.interv.mle(dat = dat, thresh = thresh,
delta = delta, rtrunc = rtrunc, ltrunc = ltrunc)
}
