% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsurv.R
\name{np_nll}
\alias{np_nll}
\title{Marginal log likelihood function of the nonparametric multinomial with censoring and truncation}
\usage{
np_nll(p, omega_lb, omega_ub, eta_lb, eta_ub, transform = TRUE)
}
\arguments{
\item{p}{vector of \code{D-1} parameters}

\item{omega_lb}{index of interval in which death occurs}

\item{omega_ub}{index of interval in which death occurs (if death is observed), or else the largest interval.}

\item{eta_lb}{vector of largest index for the lower truncation}

\item{eta_ub}{vector of smallest index for the upper truncation}

\item{transform}{logical; are parameters on logit scale? Default to \code{TRUE}}
}
\description{
Marginal log likelihood function of the nonparametric multinomial with censoring and truncation
}
