% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsurv.R
\name{np_nll_hessian}
\alias{np_nll_hessian}
\title{Hessian of (marginal), aka incomplete log likelihood function}
\usage{
np_nll_hessian(p, omega_lb, omega_ub, eta_lb, eta_ub)
}
\arguments{
\item{p}{vector of D-1 parameters}
}
\description{
Hessian of (marginal), aka incomplete log likelihood function
}
