% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pextweibull}
\alias{pextweibull}
\title{Distribution function of the extended Weibull distribution}
\usage{
pextweibull(
  q,
  scale = 1,
  shape1 = 0,
  shape2 = 1,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{vector of quantiles.}

\item{scale}{scale parameter, strictly positive.}

\item{shape1}{shape parameter of the generalized Pareto component.}

\item{shape2}{shape parameter of the Weibull component.}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}

\item{log.p}{logical; if \code{FALSE} (default), values are returned on the probability scale.}
}
\value{
a vector of (log)-probabilities of the same length as \code{q}
}
\description{
Distribution function of the extended Weibull distribution
}
\keyword{internal}
