% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pgompmake}
\alias{pgompmake}
\title{Distribution function of the Gompertz-Makeham distribution}
\usage{
pgompmake(
  q,
  scale = 1,
  shape = 1,
  lambda = 1,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{vector of quantiles.}

\item{scale}{scale parameter, strictly positive.}

\item{shape}{shape parameter.}

\item{lambda}{exponential rate}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}

\item{log.p}{logical; if \code{FALSE} (default), values are returned on the probability scale.}
}
\value{
a vector of (log)-probabilities of the same length as \code{q}
}
\description{
Distribution function of the Gompertz-Makeham distribution
}
\keyword{internal}
