% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{pgpd}
\alias{pgpd}
\title{Distribution function of the generalized Pareto distribution}
\usage{
pgpd(q, loc = 0, scale = 1, shape = 0, lower.tail = TRUE, log.p = FALSE)
}
\arguments{
\item{q}{vector of quantiles.}

\item{loc}{location parameter.}

\item{scale}{scale parameter, strictly positive.}

\item{shape}{shape parameter.}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}

\item{log.p}{logical; if \code{FALSE} (default), values are returned on the probability scale.}
}
\value{
a vector of (log)-probabilities of the same length as \code{q}
}
\description{
Distribution function of the generalized Pareto distribution
}
\keyword{internal}
