% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.elife_ecdf}
\alias{plot.elife_ecdf}
\title{Plot empirical distribution function}
\usage{
\method{plot}{elife_ecdf}(x, ...)
}
\arguments{
\item{x}{argument of class \code{elife_ecdf}}

\item{...}{additional arguments for the plot}
}
\description{
Plot empirical distribution function
}
