% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/northrop.R
\name{autoplot.elife_northropcoleman}
\alias{autoplot.elife_northropcoleman}
\alias{plot.elife_northropcoleman}
\title{P-value plot for Northrop and Coleman diagnostic}
\usage{
autoplot.elife_northropcoleman(object, ...)

\method{plot}{elife_northropcoleman}(x, plot.type = c("base", "ggplot"), plot = TRUE, ...)
}
\arguments{
\item{object}{object of class \code{elife_northropcoleman}, with the fitted piecewise-constant generalized Pareto model}

\item{...}{additional arguments for base \code{plot}}

\item{x}{an object of class \code{elife_northropcoleman}}

\item{plot.type}{string indicating the type of plot}

\item{plot}{logical; should the routine print the graph if \code{plot.type} equals \code{"ggplot"}? Default to \code{TRUE}.}
}
\value{
a base R or ggplot object with p-values for the Northrop-Coleman test against thresholds.
}
\description{
The Northrop-Coleman tests for penultimate models are
comparing the piece-wise generalized Pareto distribution
to the generalized Pareto above the lower threshold.
}
