% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endpoint.R
\name{plot.elife_profile}
\alias{plot.elife_profile}
\title{Plot profile of endpoint}
\usage{
\method{plot}{elife_profile}(x, plot.type = c("base", "ggplot"), plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{elife_profile} containing information about the profile likelihood, maximum likelihood and grid of values for the endpoint}

\item{plot.type}{string indicating whether to use base R for plots or \code{ggplot2}}

\item{plot}{logical; if \code{TRUE}, creates a plot when \code{plot.type="ggplot"}. Useful for returning \code{ggplot} objects without printing the graphs}

\item{...}{additional arguments to pass to \code{plot}, currently ignored}
}
\value{
base R or ggplot object for a plot of the profile log likelihood of the endpoint of the generalized Pareto distribution
}
\description{
Plot profile of endpoint
}
