% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/npsurv.R
\name{plot.np.surv}
\alias{plot.np.surv}
\title{Create a survival function using weights and intervals}
\usage{
\method{plot}{np.surv}(
  npi,
  conf.type = c("log", "log-log", "plain", "logit", "arcsin"),
  cumhaz = FALSE,
  ...
)
}
\arguments{
\item{npi}{output from np.interv.mle, a list with components \code{interval}, \code{par} and \code{vcov}}

\item{conf.type}{transformation for the pointwise confidence intervals;}

\item{cumhaz}{plot the cumulative hazard rather than the probability in state or survival}

\item{...}{additional arguments passed to plot}
}
\description{
Create a survival function using weights and intervals
}
