% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{pperksmake}
\alias{pperksmake}
\title{Distribution function of the Perks-Makeham distribution}
\usage{
pperksmake(
  q,
  rate = 1,
  shape = 1,
  lambda = 0,
  lower.tail = TRUE,
  log.p = FALSE
)
}
\arguments{
\item{q}{vector of quantiles.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape}{shape parameter (\eqn{\alpha})}

\item{lambda}{exponential rate of the Makeham component}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}

\item{log.p}{logical; if \code{FALSE} (default), values are returned on the probability scale.}
}
\value{
a vector of (log)-probabilities of the same length as \code{q}
}
\description{
Distribution function of the Perks-Makeham distribution
}
\keyword{internal}
