% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshstab.R
\name{prof_gp_shape_confint}
\alias{prof_gp_shape_confint}
\title{Profile log likelihood for the shape parameter of the generalized Pareto distribution}
\usage{
prof_gp_shape_confint(
  mle = NULL,
  dat,
  thresh,
  ltrunc,
  rtrunc,
  rcens,
  type = c("none", "ltrt", "ltrc"),
  level = 0.95
)
}
\arguments{
\item{mle}{an object of class \code{elife_par}}

\item{dat}{vector of threshold exceedances}

\item{thresh}{threshold}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{level}{level of the confidence interval}
}
\value{
a vector of length three with the maximum likelihood of the shape and profile-based confidence interval
}
\description{
This internal function is used to produce threshold stability plots.
}
\keyword{internal}
