% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qbeard}
\alias{qbeard}
\title{Quantile function of the Beard distribution}
\usage{
qbeard(p, rate = 1, shape1 = 1, shape2 = 1, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape1}{shape parameter (\eqn{\alpha})}

\item{shape2}{shape parameter (\eqn{\beta})}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the Beard distribution
}
\keyword{internal}
