% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{qextgp}
\alias{qextgp}
\title{Quantile function of the extended generalized Pareto distribution}
\usage{
qextgp(p, scale = 1, shape1, shape2, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{scale}{scale parameter, strictly positive.}

\item{shape1}{positive shape parameter \eqn{\beta}; model defaults to generalized Pareto when it equals zero.}

\item{shape2}{shape parameter \eqn{\gamma}; model reduces to Gompertz when \code{shape2=0}.}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the extended generalized Pareto distribution
}
\keyword{internal}
