% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qextweibull}
\alias{qextweibull}
\title{Quantile function of the extended Weibull distribution}
\usage{
qextweibull(p, scale = 1, shape1 = 0, shape2 = 1, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{scale}{scale parameter, strictly positive.}

\item{shape1}{shape parameter of the generalized Pareto component.}

\item{shape2}{shape parameter of the Weibull component.}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the extended Weibull distribution
}
\keyword{internal}
