% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qgompmake}
\alias{qgompmake}
\title{Quantile function of the Gompertz-Makeham distribution}
\usage{
qgompmake(p, scale = 1, shape = 0, lambda = 0, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{scale}{scale parameter, strictly positive.}

\item{shape}{shape parameter.}

\item{lambda}{exponential rate}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the Gompertz-Makeham distribution
}
\note{
The quantile function is defined in terms of Lambert's W function. Particular parameter combinations (small values of \code{lambda} lead to numerical overflow; the function throws a warning when this happens.
}
\keyword{internal}
