% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametric.R
\name{qgpd}
\alias{qgpd}
\title{Quantile function of the generalized Pareto distribution}
\usage{
qgpd(p, loc = 0, scale = 1, shape = 0, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{loc}{location parameter.}

\item{scale}{scale parameter, strictly positive.}

\item{shape}{shape parameter.}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the generalized Pareto distribution
}
\keyword{internal}
