% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qperks}
\alias{qperks}
\title{Quantile function of the Perks distribution}
\usage{
qperks(p, rate = 1, shape = 1, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape}{shape parameter (\eqn{\alpha})}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the Perks distribution
}
\keyword{internal}
