% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{qperksmake}
\alias{qperksmake}
\title{Quantile function of the Perks-Makeham distribution}
\usage{
qperksmake(p, rate = 1, shape = 1, lambda = 0, lower.tail = TRUE)
}
\arguments{
\item{p}{vector of probabilities.}

\item{rate}{rate parameter (\eqn{\nu})}

\item{shape}{shape parameter (\eqn{\alpha})}

\item{lambda}{exponential rate of the Makeham component}

\item{lower.tail}{logical; if \code{TRUE} (default), the lower tail probability \eqn{\Pr(X \leq x)} is returned.}
}
\value{
vector of quantiles
}
\description{
Quantile function of the Perks-Makeham distribution
}
\keyword{internal}
