% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{relife_dtrunc}
\alias{relife_dtrunc}
\title{Sample observations from a doubly truncated generalized Pareto distribution}
\usage{
relife_dtrunc(
  n,
  scale,
  shape,
  lower,
  upper,
  family = c("exp", "gp", "gomp", "weibull", "extgp")
)
}
\arguments{
\item{n}{sample size}

\item{scale}{scale parameter}

\item{shape}{shape parameter(s)}

\item{lower}{vector of lower bounds}

\item{upper}{vector of upper bounds}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}
}
\value{
a vector of \code{n} observations
}
\description{
Sample observations from a doubly truncated generalized Pareto distribution
}
