% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{test_elife}
\alias{test_elife}
\title{Likelihood ratio test for covariates}
\usage{
test_elife(
  dat,
  covariate,
  thresh,
  ltrunc = NULL,
  rtrunc = NULL,
  rcens = NULL,
  type = c("none", "ltrc", "ltrt"),
  family = c("exp", "gp", "weibull", "gomp", "extgp"),
  weights = rep(1, length(dat))
)
}
\arguments{
\item{dat}{vector of threshold exceedances}

\item{covariate}{vector of factors, logical or integer whose distinct values are}

\item{thresh}{threshold}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}

\item{weights}{weights for observations}
}
\value{
a list with elements
\itemize{
\item{\code{lrt_stat}: }{likelihood ratio statistic}
\item{\code{df}: }{degrees of freedom}
\item{\code{p_value}: }{the p-value obtained from the asymptotic chi-square approximation.}
}
}
\description{
This function fits separate models for each distinct
value of covariate and computes a likelihood ratio test
to test whether there are significante differences between
groups.
}
\examples{
with(uk110,
test_elife(dat = ndays,
thresh = 40178L, covariate = gender,
ltrunc = ltrunc, rtrunc = rtrunc,
family = "exp", type = "ltrt"))
}
