% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{uq1_qqplot_elife}
\alias{uq1_qqplot_elife}
\title{Uncertainty quantification for quantile-quantile plots}
\usage{
uq1_qqplot_elife(
  B = 9999L,
  dat,
  par,
  lower,
  upper,
  level = 0.95,
  type2 = c("none", "ltrt", "ltrc"),
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp")
)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{dat}{vector of data}

\item{par}{parameter estimates of the model}

\item{lower}{lower bounds (truncation or lowest possible value)}

\item{upper}{upper bound for right-censoring or right-truncation}

\item{level}{level of the confidence intervals}

\item{family}{string; choice of parametric family}
}
\description{
Uncertainty quantification for quantile-quantile plots
}
\keyword{internal}
