% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{uq2_qqplot_elife}
\alias{uq2_qqplot_elife}
\title{Approximate the uncertainty}
\usage{
uq2_qqplot_elife(
  dat,
  lower,
  upper,
  rcens = NULL,
  type = c("none", "ltrt", "ltrc"),
  family = c("exp", "gp", "gomp", "weibull", "extgp")
)
}
\description{
Approximate the uncertainty
}
\keyword{internal}
