% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{boxcox_transfo}
\alias{boxcox_transfo}
\title{Box-Cox transformation function}
\usage{
boxcox_transfo(par, lambda = rep(1, length(par)))
}
\description{
Given a vector of parameters, apply the Box-Cox transformation.
}
\keyword{internal}
