% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.turnbull_em}
\alias{.turnbull_em}
\title{Turnbull EM algorithm (low storage implementation)}
\usage{
.turnbull_em(
  x,
  censLow,
  censUpp,
  truncLow,
  truncUpp,
  cens = TRUE,
  trunc = TRUE,
  tol = 1e-12,
  maxiter = 100000L
)
}
\arguments{
\item{x}{vector of ordered unique elements at which to estimate probabilities}

\item{censLow}{integer vector in [0, J-1]; the lowest \code{x} value at which the observation could be observed to fail (left censoring)}

\item{censUpp}{integer vector in [0, J-1]; the largest \code{x} value at which the observation could fail (right-censoring)}

\item{truncLow}{integer vector in [0, J-1]; the lowest \code{x} such that \eqn{a_i < x} (left-truncation)}

\item{truncUpp}{integer vector in [0, J-1]; the largest \code{x} such that \eqn{b_i > x} (right-truncation)}

\item{cens}{logical; if \code{FALSE}, then \code{censUpp = censLow} and a particular update can be avoided in the EM algorithm}

\item{tol}{tolerance level for terminating the EM algorithm}

\item{maxiter}{maximum number of iteration for the EM algorithm}

\item{zerotol}{tolerance level for setting a probability to zero}
}
\value{
a list with the probabilities and the standard errors
}
\description{
Turnbull EM algorithm (low storage implementation)
}
\keyword{internal}
