% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.turnbull_intervals}
\alias{.turnbull_intervals}
\title{Turnbull's sets}
\usage{
.turnbull_intervals(Lcens, Rcens, Ltrunc, Rtrunc, status)
}
\arguments{
\item{Lcens}{set of left censoring limits}

\item{Rcens}{vector of right censoring limits}

\item{Ltrunc}{vector of left truncation limits}

\item{Rtrunc}{vector of right truncation limits}

\item{status}{integer vector giving status of censoring set}
}
\description{
Given truncation and censoring sets,
construct disjoint increasing intervals whose
left and right endpoints lie in L and R
and which contain no other members of L and R
}
\keyword{internal}
