% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard_elife}
\alias{hazard_elife}
\title{Profile likelihood for hazard}
\usage{
hazard_elife(
  par,
  dat,
  thresh = 0,
  type = c("none", "ltrt", "ltrc"),
  rcens,
  ltrunc,
  rtrunc,
  family = c("exp", "gp", "gomp", "weibull", "extgp"),
  weights = rep(1, length(dat))
)
}
\arguments{
\item{par}{vector of parameters}

\item{dat}{vector of threshold exceedances}

\item{thresh}{threshold}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}

\item{weights}{weights for observations}
}
\value{
a profile likelihood object
}
\description{
This function computes the hazard for different \code{elife} parametric
model with profile-likelihood based confidence intervals.
It is also used to provide plots with varying thresholds.
}
