% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard.R
\name{hazard_fn_elife}
\alias{hazard_fn_elife}
\title{Hazard function for various parametric models}
\usage{
hazard_fn_elife(
  x,
  par,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp")
)
}
\arguments{
\item{x}{vector of points at which to evaluate the hazard function}

\item{par}{vector of scale and shape parameters}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}), Gompertz-Makeham (\code{gompmake}) or extended generalized Pareto (\code{extgp}).}
}
\value{
a vector with the value of the hazard function at \code{x}
}
\description{
Hazard function for various parametric models
}
\keyword{internal}
