% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{ks_test}
\alias{ks_test}
\title{Goodness-of-fit diagnostics}
\usage{
ks_test(
  dat,
  thresh,
  ltrunc = NULL,
  rtrunc = NULL,
  rcens = NULL,
  type = c("none", "ltrc", "ltrt"),
  family = c("exp", "gp", "weibull", "gomp", "extgp", "gppiece"),
  B = 999L
)
}
\arguments{
\item{dat}{vector of threshold exceedances}

\item{thresh}{threshold}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}

\item{B}{number of bootstrap simulations}
}
\value{
a list with elements
\itemize{
\item{\code{stat}: }{the value of the test statistic}
\item{\code{pval}: }{p-value obtained via simulation}
}
}
\description{
Compute the Kolmogorov-Smirnov or the Anderson-Darling
test statistic and compare it with a simulated null
distribution obtained via a parametric bootstrap.
}
\note{
The bootstrap scheme requires simulating new data,
fitting a parametric model and estimating the nonparametric
maximum likelihood estimate for each new sample.
This is computationally intensive in large samples.
}
