% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot.elife_ecdf}
\alias{plot.elife_ecdf}
\title{Approximate uncertainty for diagnostic plots}
\usage{
\method{plot}{elife_ecdf}(x, ...)
}
\arguments{
\item{x}{argument of class \code{elife_ecdf}}

\item{...}{additional arguments for the plot}

\item{B}{integer; number of simulations}

\item{object}{an object of class \code{elife_par}}

\item{logscale}{logical; if \code{TRUE} (default), compute the approximation using \eqn{\log(\sigma)} rather than \eqn{\sigma}}
}
\description{
Approximate the uncertainty by resampling parameters
estimates using a normal approximation to the maximum
likelihood estimators.
}
\keyword{internal}
