% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northrop.R
\name{plot.elife_northropcoleman}
\alias{plot.elife_northropcoleman}
\title{P-value plot}
\usage{
\method{plot}{elife_northropcoleman}(x, plot.type = c("base", "ggplot"), plot = TRUE, ...)
}
\arguments{
\item{x}{an object of class \code{elife_northropcoleman}}

\item{plot.type}{string indicating the type of plot}

\item{plot}{logical; should the routine print the graph if \code{plot.type} equals \code{"ggplot"}? Default to \code{TRUE}.}

\item{...}{additional arguments for base \code{plot}}
}
\description{
The Northrop-Coleman tests for penultimate models are
comparing the piece-wise generalized Pareto distribution
to the generalized Pareto above the lower threshold.
}
