% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshstab.R
\name{prof_gp_shape_confint}
\alias{prof_gp_shape_confint}
\title{Profile log likelihood for the shape parameter of the generalized Pareto distribution}
\usage{
prof_gp_shape_confint(
  mle = NULL,
  time,
  time2 = NULL,
  event = NULL,
  thresh,
  ltrunc = NULL,
  rtrunc = NULL,
  type = c("right", "left", "interval", "interval2"),
  level = 0.95,
  weights = NULL
)
}
\arguments{
\item{mle}{an object of class \code{elife_par}}

\item{time}{excess time of the event of follow-up time, depending on the value of event}

\item{time2}{ending excess time of the interval for interval censored data only.}

\item{event}{status indicator, normally 0=alive, 1=dead. Other choices are \code{TRUE}/\code{FALSE} (\code{TRUE} for death).
For interval censored data, the status indicator is 0=right censored, 1=event at time, 2=left censored, 3=interval censored.
Although unusual, the event indicator can be omitted, in which case all subjects are assumed to have an event.}

\item{thresh}{vector of thresholds}

\item{ltrunc}{lower truncation limit, default to \code{NULL}}

\item{rtrunc}{upper truncation limit, default to \code{NULL}}

\item{type}{character string specifying the type of censoring. Possible values are "\code{right}", "\code{left}", "\code{interval}", "\code{interval2}".}

\item{level}{level of the confidence interval}

\item{weights}{weights for observations}
}
\value{
a vector of length three with the maximum likelihood of the shape and profile-based confidence interval
}
\description{
This internal function is used to produce threshold stability plots.
}
\keyword{internal}
