% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{samp_elife}
\alias{samp_elife}
\title{Simulate excess lifetime with truncation or right-censoring}
\usage{
samp_elife(
  n,
  scale,
  shape,
  lower = 0,
  upper = Inf,
  family = c("exp", "gp", "gomp", "gompmake", "weibull", "extgp", "gppiece"),
  type2 = c("none", "ltrt", "ltrc")
)
}
\arguments{
\item{n}{sample size}

\item{scale}{scale parameter(s)}

\item{shape}{shape parameter(s)}

\item{lower}{vector of lower bounds}

\item{upper}{vector of upper bounds}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}), Gompertz-Makeham (\code{gompmake}) or extended generalized Pareto (\code{extgp}).}

\item{type2}{string, either \code{none}, \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}
}
\value{
either a vector of observations or, if \code{type2=ltrc}, a list with \code{n} observations \code{dat} and a logical vector of the same length with \code{TRUE} for right-censored observations and \code{FALSE} otherwise.
}
\description{
This function dispatches simulations accounting for potential left-truncation (remove by setting lower to zero).
If \code{type2=ltrt}, simulated observations will be lower than the upper bounds \code{upper}.
If \code{type2=ltrc}, simulated observations are capped at \code{upper} and the observation is right-censored (\code{rcens=TRUE}).
}
