% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshstab.R
\name{tstab}
\alias{tstab}
\title{Threshold stability plots}
\usage{
tstab(
  dat,
  thresh,
  ltrunc,
  rtrunc,
  rcens,
  type = c("none", "ltrt", "ltrc"),
  family = c("gp", "exp"),
  method = c("wald", "profile"),
  level = 0.95,
  plot.type = c("base", "ggplot"),
  which.plot = c("scale", "shape")
)
}
\arguments{
\item{dat}{vector of threshold exceedances}

\item{thresh}{threshold}

\item{ltrunc}{lower truncation limit, possibly zero}

\item{rtrunc}{upper truncation limit}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}

\item{method}{string giving the type of pointwise confidence interval, either Wald (\code{wald}) or profile likelihood (\code{lrt})}

\item{level}{probability level for the pointwise confidence intervals}

\item{plot.type}{string; either \code{base} for base R plots or \code{ggplot} for \code{ggplot2} plots}

\item{which.plot}{string; which parameters to plot;}
}
\value{
an invisible list with pointwise estimates and confidence intervals for the scale and shape parameters
}
\description{
The generalized Pareto and exponential distribution
are threshold stable. This property, which is used
for extrapolation purposes, can also be used to diagnose
goodness-of-fit: we expect the parameters \eqn{\xi} and \eqn{\widetilde{\sigma} = \sigma + \xi u}
to be constant over a range of thresholds. The threshold stability
plot consists in plotting maximum likelihood estimates with pointwise confidence interval.
This function handles interval truncation and right-censoring.
}
\details{
The shape estimates are constrained
}
\seealso{
\code{\link[mev]{tstab.gpd}}, \code{\link[ismev]{gpd.fitrange}}, \code{\link[evd]{tcplot}}
}
