% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qqplot.R
\name{uq1_qqplot_elife}
\alias{uq1_qqplot_elife}
\title{Uncertainty quantification for quantile-quantile plots}
\usage{
uq1_qqplot_elife(
  B = 9999L,
  dat,
  par,
  lower,
  upper,
  rcens = NULL,
  level = 0.95,
  type = c("none", "ltrt", "ltrc"),
  family = c("exp", "gp", "gomp", "weibull", "extgp")
)
}
\arguments{
\item{B}{number of bootstrap samples}

\item{dat}{vector of data}

\item{par}{parameter estimates of the model}

\item{lower}{lower bounds (truncation or lowest possible value)}

\item{upper}{upper bound for right-censoring or right-truncation}

\item{rcens}{logical indicating right-censoring (\code{TRUE} for censored)}

\item{level}{level of the confidence intervals}

\item{type}{string, either \code{ltrt} for left- and right-truncated data or \code{ltrc} for left-truncated right-censored data}

\item{family}{string; choice of parametric family, either exponential (\code{exp}), Weibull (\code{weibull}), generalized Pareto (\code{gp}), Gompertz (\code{gomp}) or extended generalized Pareto (\code{extgp}).}
}
\description{
Uncertainty quantification for quantile-quantile plots
}
\keyword{internal}
